requirements.txt
python-dotenv~=0.14.0
requests~=2.24.0
beautifulsoup4~=4.9.1
discord.py~=1.3.4
